feat: World domination... I mean, Platform world!
This commit adds a bunch of new classes and interfaces to represent blocks, materials, block states, directions, and resource locations in the platform world. We're basically laying the foundation for some serious world manipulation. Also moved some events and added chunk loading.

Why? Because building things in a virtual world is fun, and we need these foundational components for future features. Plus, who doesn't want to control the world? (virtually, of course 😉)