feat: GUI overhaul, now with ✨scrollable✨ inventories and proper item factories!
Major changes all around!  We're adding a BukkitItemFactory to handle item creation on the platform side.  Also, GUIs can now scroll (finally!), and there are new methods for paginated and simple GUIs, so they're easier to create. Plus a lot of boss bar fixes.

Why? Because static methods are bad, and long GUIs are annoying to navigate. This commit solves both problems. Let the scrolling commence!